chore: move userBlockUnblockModal to modals folder
